+++ src/feregion/__init__.py
@@ -9,7 +9,7 @@ from ._default import get_default_lookup
 from .core import FlinnEngdahlLookup
 from .types import Region
 
-__version__ = "0.1.2a5"
+__version__ = "0.1.2a6"
 
 
 def lookup_number(longitude: float, latitude: float) -> int: